requirements.txt
psycopg2-binary==2.9.6
pandas==2.0.3
python-dotenv==1.0.0
sqlalchemy==2.0.20
tensorflow-recommenders==0.7.3
catboost==1.2.2
lightgbm==4.3.0
loguru==0.7.0